# docs/openapi.yaml
openapi: 3.0.0
info:
  title: Tractive API
  version: 1.0.0
  description: API documentation for Tractive backend
servers:
  - url: http://localhost:3000
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Product:
      type: object
      properties:
        _id:
          type: string
        name:
          type: string
        description:
          type: string
        price:
          type: number
        quantity:
          type: number
        discount:
          type: number
        images:
          type: array
          items:
            type: string
          description: Image URLs (Cloudinary/S3). No base64 data.
        videos:
          type: array
          items:
            type: string
          description: Video URLs (Cloudinary/S3). No base64 data.
        category:
          type: string
          nullable: true
        subcategory:
          type: string
          nullable: true
        categories:
          type: array
          items:
            type: string
          description: Normalized category list including category and subcategory.
        wishlisted:
          type: boolean
          description: True when the authenticated buyer has wishlisted the product.
        localTransport:
          type: object
          properties:
            required:
              type: boolean
            fee:
              type: number
            from:
              type: string
              nullable: true
            to:
              type: string
              nullable: true
            note:
              type: string
              nullable: true
...
    Transaction:
      type: object
      properties:
        order:
          type: string
        buyer:
          type: string
        amount:
          type: number
        status:
          type: string
security:
  - bearerAuth: []
paths:
...
  /api/products:
    get:
      summary: List products
      description: Public list endpoint. If authenticated with activeRole=agent and owner is not provided, results are scoped to that agent's own products.
      parameters:
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: status
          schema:
            type: string
            enum: [available, out_of_stock, discontinued]
        - in: query
          name: category
          schema:
            type: string
        - in: query
          name: subcategory
          schema:
            type: string
        - in: query
          name: farmer
          schema:
            type: string
        - in: query
          name: owner
          schema:
            type: string
        - in: query
          name: minPrice
          schema:
            type: number
        - in: query
          name: maxPrice
          schema:
            type: number
        - in: query
          name: from
          schema:
            type: string
            format: date-time
        - in: query
          name: to
          schema:
            type: string
            format: date-time
        - in: query
          name: sortBy
          schema:
            type: string
        - in: query
          name: sortOrder
          schema:
            type: string
            enum: [asc, desc]
        - in: query
          name: full
          schema:
            type: boolean
          description: Return full product documents (default is compact list)
        - in: query
          name: includeMedia
          schema:
            type: boolean
          description: Include images/videos in compact list (default true)
      responses:
        '200':
          description: Paginated list of products
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Product'
                  products:
                    type: array
                    items:
                      $ref: '#/components/schemas/Product'
                  pagination:
                    type: object
                    properties:
                      page:
                        type: integer
                      limit:
                        type: integer
                      total:
                        type: integer
    post:
      summary: Create product
      description: Requires `activeRole` to be `agent` or `admin`. Preferred category shape is `category` + `subcategory` + `categories`, while `localTransport` describes seller-side farm-to-pickup movement that is later added into buyer checkout totals.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
            example:
              name: "Fresh Maize"
              description: "High quality yellow maize from Kaduna farms"
              price: 3800
              quantity: 100
              discount: 10
              unit: "kg"
              category: "Grains"
              subcategory: "Maize"
              categories: ["Grains", "Maize"]
              images: ["https://example.com/maize1.jpg"]
              videos: ["https://example.com/maize1.mp4"]
              farmer: "507f1f77bcf86cd799439012"
              localTransport:
                required: true
                fee: 2500
                from: "Kachia Farm"
                to: "Kaduna Aggregation Point"
                note: "Farm gate pickup to interstate loading point"
      responses:
        '201':
          description: Product created
          content:
            application/json:
              example:
                message: "Product created successfully"
                product:
                  _id: "67deprod1234567890abcd1"
                  name: "Fresh Maize"
                  description: "High quality yellow maize from Kaduna farms"
                  price: 3800
                  quantity: 100
                  discount: 10
                  unit: "kg"
                  owner: "67deagent1234567890abc1"
                  farmer: "507f1f77bcf86cd799439012"
                  images: ["https://example.com/maize1.jpg"]
                  videos: ["https://example.com/maize1.mp4"]
                  status: "available"
                  category: "Grains"
                  subcategory: "Maize"
                  categories: ["Grains", "Maize"]
                  localTransport:
                    required: true
                    fee: 2500
                    from: "Kachia Farm"
                    to: "Kaduna Aggregation Point"
                    note: "Farm gate pickup to interstate loading point"
...
  /api/transactions:
    get:
      summary: List your transactions
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        '200':
          description: List of transactions
    post:
      summary: Create transaction for paid order
      description: Use this after creating an order. This endpoint records the buyer payment intent for an order. It requires `activeRole` `buyer`, requires `amount` to exactly match `order.totalAmount`, and reuses any existing pending or approved transaction for the same order instead of creating duplicates.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Transaction'
            example:
              order: "507f1f77bcf86cd799439014"
              amount: 50000
              paymentMethod: "bank_transfer"
      responses:
        '201':
          description: Transaction created
  /api/transactions/{id}:
    get:
      summary: Get transaction by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Transaction details
    patch:
      summary: Legacy transaction update (admin only)
      description: Legacy endpoint. Prefer PATCH /api/transactions/{id}/status for status updates.
      deprecated: true
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [approved]
            example:
              status: "approved"
      responses:
        '200':
          description: Transaction approved
...